feat: move test IDs to common configs
