+++ packages/sage-react/lib/configs/automationTestIds/index.js
@@ -1,0 +1,1 @@
+export { TestIds } from './automationTestIds';
