+++ packages/sage-react/lib/configs/index.js
@@ -1,3 +1,4 @@
 export { SageClassnames } from './classnames';
 export { SageDictionary } from './dictionary';
 export { SageTokens } from './tokens';
+export { TestIds } from './automationTestIds';
