+++ packages/sage-react/lib/index.js
@@ -93,4 +93,3 @@ export {
   useSageBreakpoint,
   SageBreakpointProvider,
 } from './common/providers';
-export { TestIds } from './automationTestIds';
